requirements.txt
fastapi
uvicorn
sqlalchemy
stripe
bcrypt
python-dotenv
PyJWT